Scope mapEntry variable to for..of loop in TemplateWriter
Fixes #722. Previously we wouldn't scope the value of the mapEntry
variable to each iteration when looping over the templates to write.
This became an issue of any of the generated promises threw an error,
since the error message would always report that it stemmed from the
template from the last iteration.