Remove parse-filepath dependency
As described in #1903 this commit removes the parse-filepath dependency from 11ty in favor of the native path.parse.

Some files needed renaming of some variables which were named "path" and shadowed the path module. I hope I selected reasonable replacements.

In TemplatePath.js I also changed the documentation links to point to the nodejs api doc instead of the npm package.

According to a local test run it does not change test results.

Signed-off-by: Raphael Höser <raphael@hoeser.info>